Spagettied code together
Did my best to put the code together in a way that makes sense though there are some things that dont work just yet.

The purpose of FancySpagetti is so that it's the centralized hub for where we pull out changes each week (after the meetings and all that). THat way, we'll all be caught up on different changes in a slightly easier way.

Also, the only reason it's called FancySpagetti is because the code is still spagetti. It's just fancier now.