Added Item Spawner
Deleted the cube on the map and added an Item Spawner. It spawns the different pickups within a 20 unit radius (The variable is public so that it could be tweaked if needed). The actual item spawning is currently a WIP as some items could spawn under the map due to how Random.insideUnitSphere works.